+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.4.3
+
 internal:
 
 * rename CHANGES -> CHANGELOG for https://yarnpkg.com/ to display changes